Changed data type of field

+++ Settings.cs
@@ -8,7 +8,7 @@ internal static class Settings
     internal static readonly int MinPlayers = 1;
     internal static readonly int MaxPlayers = 7;
 
-    internal static readonly decimal InitialWinnings = 100m;
+    internal static readonly int InitialWinnings = 100;
 
     internal static readonly int MinBet = 5;
     internal static readonly int MaxBet = 50;
